# openapi.yaml
openapi: 3.1.0
info:
  title: agentic-saffron-api
  description: 'AI-powered institutional intelligence platform: ingestion, canonical
    data, policy-bound agents, and internet intelligence'
  version: 0.1.0
paths:
...
  /v1/intelligence/profile:
    get:
      tags:
      - intelligence
      summary: The institution's intelligence profile
      operationId: get_profile_v1_intelligence_profile_get
      parameters:
      - name: institution_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Profile V1 Intelligence Profile Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - intelligence
      summary: Create or update the intelligence profile
      operationId: put_profile_v1_intelligence_profile_put
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Put Profile V1 Intelligence Profile Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/intelligence/investigate:
    post:
      tags:
      - intelligence
      summary: Search public sources about the institution now
      operationId: investigate_v1_intelligence_investigate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvestigateBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Investigate V1 Intelligence Investigate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/intelligence/mentions:
    get:
      tags:
      - intelligence
      summary: Stored public mentions (evidence records)
      operationId: mentions_v1_intelligence_mentions_get
      parameters:
      - name: institution_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Id
      - name: days
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Days
      - name: source_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Type
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: kept
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Mentions V1 Intelligence Mentions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/intelligence/digest:
    get:
      tags:
      - intelligence
      summary: Daily intelligence digest from continuous monitoring
      operationId: digest_v1_intelligence_digest_get
      parameters:
      - name: institution_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Id
      - name: days
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Days
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Digest V1 Intelligence Digest Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/intelligence/monitor/run:
    post:
      tags:
      - intelligence
      summary: Run a monitoring pass now (normally scheduled)
      operationId: run_monitor_v1_intelligence_monitor_run_post
      parameters:
      - name: institution_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Id
      - name: background
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Background
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Run Monitor V1 Intelligence Monitor Run Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/intelligence/reports:
    get:
      tags:
      - intelligence
      summary: Past investigation reports
      operationId: reports_v1_intelligence_reports_get
      parameters:
      - name: institution_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Reports V1 Intelligence Reports Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/reports:
    get:
      tags:
      - platform
      summary: Generated reports
      operationId: list_reports_v1_reports_get
      parameters:
      - name: institution_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Institution Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Reports V1 Reports Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InvestigateBody:
      properties:
        question:
          anyOf:
          - type: string
            maxLength: 300
          - type: 'null'
          title: Question
        window_days:
          type: integer
          maximum: 365.0
          minimum: 1.0
          title: Window Days
          default: 7
        topics:
          anyOf:
          - items:
              type: string
            type: array
            maxItems: 10
          - type: 'null'
          title: Topics
        max_results:
          type: integer
          maximum: 30.0
          minimum: 1.0
          title: Max Results
          default: 10
        institution_id:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Institution Id
      type: object
      title: InvestigateBody
...
    ProfileBody:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 2
          title: Name
        location:
          type: string
          maxLength: 120
          title: Location
          default: ''
        aliases:
          items:
            type: string
          type: array
          maxItems: 50
          title: Aliases
        official_domains:
          items:
            type: string
          type: array
          maxItems: 20
          title: Official Domains
        programs:
          items:
            type: string
          type: array
          maxItems: 50
          title: Programs
        social_accounts:
          items:
            type: string
          type: array
          maxItems: 20
          title: Social Accounts
        keywords:
          items:
            type: string
          type: array
          maxItems: 20
          title: Keywords
        exclusions:
          items:
            type: string
          type: array
          maxItems: 20
          title: Exclusions
        monitoring_enabled:
          type: boolean
          title: Monitoring Enabled
          default: false
        alert_recipients:
          items:
            type: string
          type: array
          maxItems: 20
          title: Alert Recipients
        security_contacts:
          items:
            type: string
          type: array
          maxItems: 10
          title: Security Contacts
          description: Email addresses of whoever runs the institution's sites; high-severity
            incidents are emailed to them at once
        institution_id:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Institution Id
      type: object
      required:
      - name
      title: ProfileBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...